backend/app/models/ai_model.py
"""
Modelo de IA simulado para análisis de imágenes de piel
Este módulo simula el comportamiento de un modelo real de machine learning
"""

import random
import time
import asyncio
import logging
from typing import Dict, Tuple, Optional
from PIL import Image
import numpy as np
from datetime import datetime

logger = logging.getLogger(__name__)

class SkinCancerAIModel:
    """
    Simulador de modelo de IA para detección de cáncer de piel
    """
    
    def __init__(self):
        self.model_version = "1.0.0-simulated"
        self.training_date = "2024-01-01"
        self.accuracy = 0.95  # Precisión simulada del modelo
        self.is_loaded = False
        self.load_model()
    
    def load_model(self):
        """Simular carga del modelo"""
        logger.info("Cargando modelo de IA simulado...")
        time.sleep(0.1)  # Simular tiempo de carga
        self.is_loaded = True
        logger.info("Modelo de IA cargado exitosamente")
    
    def preprocess_image(self, image_path: str) -> Dict:
        """
        Preprocesar imagen para análisis
        En un modelo real, esto incluiría normalización, redimensionado, etc.
        """
        try:
            with Image.open(image_path) as img:
                # Convertir a RGB si es necesario
                if img.mode != 'RGB':
                    img = img.convert('RGB')
                
                width, height = img.size
                
                # Simular extracción de características
                features = {
                    'dimensions': (width, height),
                    'aspect_ratio': width / height,
                    'total_pixels': width * height,
                    'color_channels': len(img.getbands()),
                    'format': img.format
                }
                
                # Simular análisis de color dominante
                colors = img.getcolors(maxcolors=256*256*256)
                if colors:
                    dominant_color = max(colors, key=lambda x: x[0])[1]
                    features['dominant_color'] = dominant_color
                
                return features
                
        except Exception as e:
            logger.error(f"Error en preprocesamiento: {str(e)}")
            raise ValueError(f"Error procesando imagen: {str(e)}")
    
    def extract_medical_features(self, features: Dict) -> Dict:
        """
        Simular extracción de características médicas relevantes
        En un modelo real, esto analizaría patrones específicos de la piel
        """
        width, height = features['dimensions']
        
        # Simular características médicas basadas en la imagen
        medical_features = {
            'asymmetry_score': random.uniform(0.1, 0.9),
            'border_irregularity': random.uniform(0.0, 0.8),
            'color_variation': random.uniform(0.2, 0.7),
            'diameter_factor': min(width, height) / 100,  # Factor basado en tamaño
            'texture_complexity': random.uniform(0.3, 0.9),
            'contrast_level': random.uniform(0.4, 0.8)
        }
        
        return medical_features
    
    def calculate_risk_probability(self, medical_features: Dict, image_quality: float) -> Tuple[float, float]:
        """
        Calcular probabilidad de riesgo basada en características médicas
        Retorna (probabilidad, confianza)
        """
        # Pesos simulados para diferentes características (basados en criterios ABCDE)
        weights = {
            'asymmetry_score': 0.25,
            'border_irregularity': 0.20,
            'color_variation': 0.20,
            'diameter_factor': 0.15,
            'texture_complexity': 0.10,
            'contrast_level': 0.10
        }
        
        # Calcular score ponderado
        weighted_score = 0
        for feature, value in medical_features.items():
            if feature in weights:
                weighted_score += value * weights[feature]
        
        # Convertir score a probabilidad (0-100%)
        base_probability = weighted_score * 100
        
        # Añadir variabilidad realista
        noise = random.uniform(-10, 10)
        probability = max(5, min(85, base_probability + noise))
        
        # Calcular confianza basada en calidad de imagen y características
        base_confidence = 0.80
        quality_bonus = image_quality * 0.15
        feature_consistency = 1 - abs(0.5 - weighted_score)  # Más confianza en extremos
        
        confidence = min(0.95, base_confidence + quality_bonus + (feature_consistency * 0.05))
        
        return round(probability, 1), round(confidence, 2)
    
    def assess_image_quality(self, features: Dict) -> float:
        """
        Evaluar calidad de la imagen para el análisis médico
        Retorna score de 0.0 a 1.0
        """
        width, height = features['dimensions']
        
        # Factores de calidad
        resolution_score = min(1.0, (width * height) / (1024 * 1024))  # Mejor con más resolución
        aspect_ratio = features['aspect_ratio']
        aspect_score = 1.0 - abs(1.0 - aspect_ratio) * 0.5  # Mejor si es cuadrada
        
        # Score combinado
        quality_score = (resolution_score * 0.7) + (aspect_score * 0.3)
        
        return min(1.0, max(0.3, quality_score))  # Entre 0.3 y 1.0
    
    async def analyze_image(self, image_path: str) -> Dict:
        """
        Análisis principal de la imagen
        """
        if not self.is_loaded:
            raise RuntimeError("Modelo no está cargado")
        
        start_time = time.time()
        
        try:
            # 1. Preprocesar imagen
            logger.info("Preprocesando imagen...")
            features = self.preprocess_image(image_path)
            
            # Simular tiempo de procesamiento
            await asyncio.sleep(random.uniform(0.5, 1.5))
            
            # 2. Evaluar calidad de imagen
            image_quality = self.assess_image_quality(features)
            
            # 3. Extraer características médicas
            logger.info("Extrayendo características médicas...")
            medical_features = self.extract_medical_features(features)
            
            # Simular más tiempo de procesamiento
            await asyncio.sleep(random.uniform(1.0, 2.0))
            
            # 4. Calcular probabilidad de riesgo
            logger.info("Calculando probabilidad de riesgo...")
            probability, confidence = self.calculate_risk_probability(medical_features, image_quality)
            
            processing_time = time.time() - start_time
            
            # 5. Generar resultado
            result = {
                'probability': probability,
                'confidence': confidence,
                'processing_time': round(processing_time, 2),
                'image_quality_score': round(image_quality, 2),
                'model_version': self.model_version,
                'analysis_timestamp': datetime.now().isoformat(),
                'features_analyzed': len(medical_features),
                'medical_features': medical_features,  # Para debugging (no enviar en producción)
                'recommendations': self._generate_recommendations(probability, confidence, image_quality)
            }
            
            logger.info(f"Análisis completado: {probability}% riesgo, {confidence} confianza")
            return result
            
        except Exception as e:
            logger.error(f"Error en análisis de IA: {str(e)}", exc_info=True)
            raise
    
    def _generate_recommendations(self, probability: float, confidence: float, image_quality: float) -> Dict:
        """
        Generar recomendaciones basadas en el resultado
        """
        recommendations = {
            'urgency_level': 'low',
            'follow_up_recommended': False,
            'specialist_consultation': False,
            'monitoring_frequency': 'annual',
            'additional_tests': []
        }
        
        if probability <= 30:
            recommendations.update({
                'urgency_level': 'low',
                'follow_up_recommended': True,
                'monitoring_frequency': 'annual',
                'message': 'Riesgo bajo detectado. Continúe con chequeos regulares.'
            })
        elif probability <= 60:
            recommendations.update({
                'urgency_level': 'moderate',
                'follow_up_recommended': True,
                'specialist_consultation': True,
                'monitoring_frequency': 'semi-annual',
                'message': 'Riesgo moderado. Se recomienda consulta con dermatólogo.'
            })
        else:
            recommendations.update({
                'urgency_level': 'high',
                'follow_up_recommended': True,
                'specialist_consultation': True,
                'monitoring_frequency': 'quarterly',
                'additional_tests': ['dermoscopy', 'biopsy_consideration'],
                'message': 'Riesgo alto detectado. Consulte con un dermatólogo lo antes posible.'
            })
        
        # Ajustar por confianza
        if confidence < 0.7:
            recommendations['message'] += ' Nota: Confianza del análisis es baja, se recomienda repetir con mejor calidad de imagen.'
        
        # Ajustar por calidad de imagen
        if image_quality < 0.5:
            recommendations['image_quality_warning'] = 'La calidad de la imagen puede afectar la precisión del análisis.'
        
        return recommendations
    
    def get_model_info(self) -> Dict:
        """
        Información del modelo
        """
        return {
            'model_name': 'SkinCancer AI Detector',
            'version': self.model_version,
            'training_date': self.training_date,
            'accuracy': self.accuracy,
            'is_loaded': self.is_loaded,
            'supported_formats': ['JPEG', 'PNG'],
            'analysis_criteria': ['Asymmetry', 'Border', 'Color', 'Diameter', 'Texture'],
            'confidence_range': '0.70 - 0.95',
            'processing_time_range': '2-5 seconds'
        }

# Instancia global del modelo
ai_model = SkinCancerAIModel()